RestauranteCodenation.Application/App/UsuarioAplicacao.cs
using Microsoft.AspNetCore.Identity;
using Microsoft.Extensions.Options;
using Microsoft.IdentityModel.Tokens;
using RestauranteCodenation.Application.Interface;
using RestauranteCodenation.Application.ViewModel;
using RestauranteCodenation.Domain.Repositorio;
using System;
using System.IdentityModel.Tokens.Jwt;
using System.Text;
using System.Threading.Tasks;

namespace RestauranteCodenation.Application.App
{
    public class UsuarioAplicacao : IUsuarioAplicacao
    {
        private readonly IUsuarioRepositorio _repo;
        private readonly Token _token;
        public UsuarioAplicacao(IUsuarioRepositorio repo,
            IOptions<Token> token)
        {
            _repo = repo;
            _token = token?.Value;
        }

        public async Task<bool> Incluir(UsuarioViewModel usuario)
        {
            return await _repo.Incluir(usuario.Nome, usuario.Email, usuario.Senha);
        }

        public async Task<string> Login(LoginViewModel login)
        {
            var usuario = await _repo.Login(login.Email, login.Senha);
            return GerarToken(usuario);
        }

        private string GerarToken(IdentityUser usuario)
        {
            if (usuario == null) return null;
            var tokenHandler = new JwtSecurityTokenHandler();
            var key = Encoding.ASCII.GetBytes(_token.Secret);

            var descriptor = new SecurityTokenDescriptor
            {
                Issuer = _token.Emissor,
                Audience = _token.ValidoEm,
                Expires = DateTime.UtcNow.AddHours(_token.ExpiracaoHoras),
                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
            };

            var token = tokenHandler.CreateToken(descriptor);

            return tokenHandler.WriteToken(token);

        }
    }
}
